Added proper nested routes support for school profile
